feat: 데스크탑 Global Header 마크업
