+++ src/components/GlobalHeader/css/GlobalHeader.module.css
@@ -1,0 +1,36 @@
+.header {
+  width: 100%;
+  height: 3rem;
+  border-bottom: 1px solid var(--color-border);
+}
+
+.container {
+  position: relative;
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  height: 100%;
+  padding: 0 var(--side-padding);
+  margin: 0 auto;
+}
+
+.title {
+  position: absolute;
+  top: 50%;
+  left: 50%;
+  font-size: var(--font-18);
+  transform: translate(-50%, -50%);
+}
+
+@media screen and (min-width: 48rem) {
+  .container {
+    max-width: var(--screen-max-width);
+  }
+
+  .title {
+    position: relative;
+    top: 0;
+    left: 0;
+    transform: none;
+  }
+}
