+++ src/components/css/SearchForm.module.css
@@ -1,0 +1,23 @@
+.form {
+  position: relative;
+  overflow: hidden;
+}
+
+.search {
+  padding: 0.25rem 2.5rem 0.25rem 1rem;
+  font-size: var(--font-14);
+  border: 1px solid var(--color-border-dark);
+  border-radius: 1rem;
+}
+
+.search::placeholder {
+  font-size: var(--font-14);
+}
+
+.button {
+  position: absolute;
+  top: 50%;
+  right: 0.1rem;
+  padding: 0.5rem 0.75rem;
+  transform: translateY(-50%);
+}
